fix: StakeModal chainId issue
