+++ packages/uikit/src/widgets/Pool/StakeModal.tsx
@@ -9,7 +9,6 @@ import removeTrailingZeros from "@pancakeswap/utils/removeTrailingZeros";
 
 import PercentageButton from "./PercentageButton";
 import getThemeValue from "../../util/getThemeValue";
-import ZkSyncWarning from "./ZkSyncWarning";
 
 import {
   Box,
@@ -70,6 +69,7 @@ interface StakeModalProps {
   handleConfirmClick: any;
   pendingTx: boolean;
   imageUrl?: string;
+  warningComponent?: React.ReactElement;
 }
 
 export const StakeModal: React.FC<React.PropsWithChildren<StakeModalProps>> = ({
@@ -95,6 +95,7 @@ export const StakeModal: React.FC<React.PropsWithChildren<StakeModalProps>> = ({
   pendingTx,
   handleConfirmClick,
   imageUrl = "/images/tokens/",
+  warningComponent,
 }) => {
   const { t } = useTranslation();
   const theme = useTheme();
@@ -271,7 +272,7 @@ export const StakeModal: React.FC<React.PropsWithChildren<StakeModalProps>> = ({
           <PercentageButton onClick={() => handleChangePercent(75)}>75%</PercentageButton>
           <PercentageButton onClick={() => handleChangePercent(100)}>{t("Max")}</PercentageButton>
         </Flex>
-        <ZkSyncWarning />
+        {warningComponent}
         {!isRemovingStake && (
           <Flex mt="24px" alignItems="center" justifyContent="space-between">
             <Text mr="8px" color="textSubtle">
